Phonology modules per language are structured
