+++ src/cltk/phonology/lat/phonology.py
@@ -1,0 +1,36 @@
+"""
+
+"""
+
+import cltk.phonology.lat.transcription as latt
+
+__author__ = ["Clément Besnier <clem@clementbesnier.fr>"]
+
+
+class LatinTranscription:
+    def __init__(self):
+        self.transcription = latt.Transcriber("Classical", "Allen")
+
+    def transcribe(self, word):
+        return self.transcription.transcribe(word)
+
+    def __repr__(self):
+        return f"<LatinTranscription>"
+
+    def __call__(self, word):
+        return self.transcribe(word)
+
+
+class LatinSyllabifier:
+    def __init__(self):
+        self.transcription = latt.Transcriber("Classical", "Allen")
+
+    def syllabify(self, word):
+        return latt.Word(self.transcription.transcribe(word),
+                         latt.LATIN["Classical"]["Allen"]).syllabify()
+
+    def __repr__(self):
+        return f"<LatinSyllabifier>"
+
+    def __call__(self, word):
+        return self.syllabify(word)
